klientctl: remove codegangsta/cli package from kd logic
