+++ go/src/koding/klientctl/commands/config/cmd.go
@@ -1,10 +1,38 @@
 package config
 
 import (
+	"os"
+	"path/filepath"
+
+	konfig "koding/kites/config"
 	"koding/klientctl/commands/cli"
 
 	"github.com/spf13/cobra"
 )
+
+var (
+	// ConfigFolder is folder where config and other related info are stored.
+	ConfigFolder string
+)
+
+func init() {
+	var err error
+	if ConfigFolder, err = createFolderAtHome(".config", "koding"); err != nil {
+		panic(err)
+	}
+}
+
+func createFolderAtHome(cf ...string) (string, error) {
+	args := []string{konfig.CurrentUser.HomeDir}
+	args = append(args, cf...)
+
+	folderName := filepath.Join(args...)
+	if err := os.MkdirAll(folderName, 0755); err != nil {
+		return "", err
+	}
+
+	return folderName, nil
+}
 
 // NewCommand creates a command that manages KD configuration.
 func NewCommand(c *cli.CLI) *cobra.Command {
