--- go/src/koding/klientctl/config_darwin.go
@@ -1,6 +1,0 @@
-package main
-
-const (
-	// The location of the kd logfile.
-	LogFilePath = "/Library/Logs/kd.log"
-)
