--- go/src/koding/klientctl/config_linux.go
@@ -1,8 +1,0 @@
-package main
-
-const (
-	// The location of the kd logfile.
-	//
-	// TODO: We may need multiple, for different OSs?
-	LogFilePath = "/var/log/kd.log"
-)
